Refactor: move entity interfaces
