+++ src/hooks/useGenres.ts
@@ -2,14 +2,9 @@ import { useQuery } from "@tanstack/react-query";
 import ms from "ms";
 import genres from "../data/genres";
 import APIClient from "../services/api-client";
+import { Genre } from "../entities/Genre";
 
 const apiClient = new APIClient<Genre>("/genres");
-export interface Genre {
-  id: number;
-  name: string;
-  image_background: string;
-}
-
 const useGenres = () =>
   useQuery({
     queryKey: ["genres"],
